requirements.txt
# requirements.txt (optional, not used by pip install .)
pyspark==2.4.4
pandas>=1.0.0
numpy>=1.18.0
scipy>=1.5.0
matplotlib>=3.0.0
pyyaml>=5.0.0